requirements.txt
gradio
openai
azure-cognitiveservices-speech
python-dotenv
boto3